Changed `Print' to `Query', N before M
The equivalent sets program's initialization function had N and M the
wrong way around. Also, I forgot to change the enum for query modes so
that it uses the word `QUERY' instead.

+++ eqSets.h
@@ -8,7 +8,7 @@
 #include <stdlib.h>
 
 // Reconfigure Equivalent Set Generation
-int eqSetsInit(unsigned long, size_t);
+int eqSetsInit(size_t, unsigned long);
 
 // Enumerate Equivalent Sets
 int eqSets(const unsigned long *, size_t,
